+++ mmvm_v2_tests/guest_vm/tests/standalone_alternate_program.js
@@ -1,1 +1,7 @@
-print("Alternate standalone workload");
+if (arguments[0] === "--environment") {
+    print(String(process.env.DISPLAY) + "|" +
+          String(process.env.XAUTHORITY) + "|" +
+          String(process.env.HOME));
+} else {
+    print("Alternate standalone workload");
+}